+++ example/my-ext/pogo/components/timer.scss
