+++ example/my-ext/pogo/server/pogo.server.ts
@@ -2,8 +2,8 @@ import { z } from 'zod';
 import _ from 'lodash';
 import fetch from 'node-fetch';
 import { getDB, withDb } from './db.server';
-import { getMegaPokemon, getRegisteredPokemon } from './pogoapi.server';
-import { Species } from './domain-utils';
+import { getMegaPokemon } from './pogoapi.server';
+import { Species } from '../domain-utils';
 
 // Going to start by making a mega evolution planner.
